"ads2_m6_a1_1: 0/12 testcases passed. Cannot run hidden testcases as Server is unreachablecheck_style score: 0/1 "

+++ ADS - 2 Assignments/m6/ADS-2Week-1/Solution.java
@@ -19,7 +19,7 @@ class PageRank {
 				// ranks[j] = ((ranks[j])/(digraph.outdegree(j)));
 				temp = 0.0;
 				for (int k : revdigraph.adj(j)){
-					temp += ((ranks[k])/(digraph.outdegree(k)));
+					temp += ((ranks[k])/((double)digraph.outdegree(k)));
 					System.out.println(j);
 				}
 				ranks[j] = temp;
